tr2/phase/cutscene: extract cutscene module
