+++ src/tr2/meson.build
@@ -92,6 +92,7 @@ sources = [
   'game/console/common.c',
   'game/console/setup.c',
   'game/creature.c',
+  'game/cutscene.c',
   'game/demo.c',
   'game/effects.c',
   'game/fmv.c',
